gema-rag/src/requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.0
torch==2.1.0
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.1.2
langdetect==1.0.9
spacy==3.7.2
rank-bm25==0.2.2
neo4j==5.14.1
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0
requests==2.31.0
groq==0.4.0
python-dotenv==1.0.0
numpy==1.24.3
tqdm==4.66.1
